requirements.txt
bjoern
requests==2.32.0
Flask
opentelemetry-api
opentelemetry-sdk
opentelemetry-instrumentation
opentelemetry-instrumentation-flask
opentelemetry-exporter-otlp